Implemented new orientation model syntax and parser.
Instead of a plaintext file, orientation models are now represented
using .orientation files. These files are recognised by an
Xtext-generated plugin for Eclipse (in a separate project) which makes
writing valid orientation models easier.

In order to use the new format, I have created the OrientationParser
class which uses an Xtext-generated parser for the model. This extracts
EMF objects, from which I have created my own representations.

The orientation stamper was updated to use information from the new
orientation model format. This required a serializable representation of
a model, and is one of the reasons I chose to use my own representation
instead of the EMF objects.

Finally, the existing tests were updated to use the new orientation
model format.